+++ R/data.R
@@ -1,12 +1,13 @@
 #' @title Example Purple Air Synoptic dataset
-#' @format A tibble with 10081 rows and 42 columns of data.
+#' @format A tibble with 10220 rows and 45 columns of data.
 #' @description The \code{example_pas} dataset provides a quickly loadable
 #' version of a \emph{pa_synoptic} object for practicing and code examples.
-#' This dataset was generatedon 2019-12-23 by running:
+#' This dataset was generatedon 2020-01-03 by running:
 #'
 #' \preformatted{
 #'   initializeMazamaSpatialUtils()
 #'   example_pas <- pas_createNew()
+#'   save(example_pas, file = "data/example_pas.rda")
 #' }
 #'
 #' @seealso example_pas_raw
@@ -19,12 +20,13 @@
 #' @description The \code{example_pas_raw} dataset provides a quickly loadable
 #' version of raw Purple Air synoptic data JSON for practicing and code
 #' examples. This dataset contains data for sensors in Washington and Oregon
-#' and was generated on 2019-12-23 by running:
+#' and was generated on 2020-01-03 by running:
 #' 
 #' \preformatted{
 #'   example_pas_raw <- 
 #'     downloadParseSynopticData() %>%
 #'     dplyr::filter(Lon > -125.0 & Lon < -117.0 & Lat > 42.0 & Lat < 49.0)
+#'   save(example_pas_raw, file = "data/example_pas_raw.rda")
 #' }
 #' 
 #' This dataset can be converted into a standard \emph{pas} dataset with:
@@ -42,13 +44,17 @@
 #' @format An S3 object composed of "meta" and "data" data.
 #' @description The \code{example_pat} dataset provides a quickly loadable version of
 #' a \emph{pa_timeseries} object for practicing and code examples.
-#' This dataset was was generated on 2019-12-30 by running:
+#' This dataset was was generated on 2020-01-03 by running:
 #'
 #' \preformatted{
-#'   example_pat <- pat_createNew(example_pas,
-#'                                label = "Seattle",
-#'                                startdate = "2018-08-01",
-#'                                enddate = "2018-08-28")
+#'   example_pat <-
+#'     pat_createNew(
+#'       label = "Seattle",
+#'       pas = example_pas,
+#'       startdate = "2018-08-01",
+#'       enddate = "2018-08-28"
+#'     )
+#'   save(example_pat, file = "data/example_pat.rda")
 #' }
 #' @seealso example_pat_failure_A
 #' @seealso example_pat_failure_B
@@ -59,14 +65,17 @@
 #' @format An S3 object composed of "meta" and "data" data.
 #' @description The \code{example_pat_failure_A} dataset provides a quickly loadable
 #' version of a \emph{pa_timeseries} object for practicing and code examples.
-#' This dataset was was generated on 2019-12-30 by running:
+#' This dataset was was generated on 2020-01-03 by running:
 #'
 #' \preformatted{
 #'   example_pat_failure_A <-
-#'     pat_createNew(example_pas,
-#'                   label = "SCNP_20",
-#'                   startdate = "2019-04-01",
-#'                   enddate = "2019-04-18")
+#'     pat_createNew(
+#'       label = "SCNP_20",
+#'       pas = example_pas,
+#'       startdate = "2019-04-01",
+#'       enddate = "2019-04-18"
+#'     )
+#'   save(example_pat_failure_A, file = "data/example_pat_failure_A.rda")
 #' }
 #' @seealso example_pat
 #' @seealso example_pat_failure_B
@@ -77,37 +86,43 @@
 #' @format An S3 object composed of "meta" and "data" data.
 #' @description The \code{example_pat_failure_B} dataset provides a quickly loadable
 #' version of a \emph{pa_timeseries} object for practicing and code examples.
-#' This dataset was was generated on 2019-12-30 by running:
+#' This dataset was was generated on 2020-01-03 by running:
 #'
 #' \preformatted{
 #'   example_pat_failure_B <-
-#'     pat_createNew(example_pas, 
-#'                   label = "SCTV_16",
-#'                   startdate = "2019-06-01",
-#'                   enddate = "2019-06-18")
+#'     pat_createNew(
+#'       label = "SCTV_16",
+#'       pas = example_pas,
+#'       startdate = "2019-06-01",
+#'       enddate = "2019-06-18"
+#'     )
+#'   save(example_pat_failure_B, file = "data/example_pat_failure_B.rda")
 #' }
 #' @seealso example_pat
 #' @seealso example_pat_failure_A
 "example_pat_failure_B"
 
 
-#' #' @title Example AirSensor Timeseries dataset
-#' #' @format An S3 object composed of "meta" and "data" data.
-#' #' @description The \code{example_sensor} dataset provides a quickly loadable version of
-#' #' an \emph{airsensor} object for practicing and code examples.
-#' #' This dataset was was generated on 2019-07-06 by running:
-#' #' 
-#' #' \preformatted{
-#' #'   pas <- pas_load()
-#' #'   example_sensor <- 
-#' #'     pat_createNew(pas, label = "SCAN_14",
-#' #'                   startdate = "2018-08-14",
-#' #'                   enddate = "2018-09-07") %>%
-#' #'     pat_createAirSensor(period = "1 hour")
-#' #' }
-#' "example_sensor"
-#' 
-#' 
+#' @title Example AirSensor Timeseries dataset
+#' @format An S3 object composed of "meta" and "data" data.
+#' @description The \code{example_sensor} dataset provides a quickly loadable version of
+#' an \emph{airsensor} object for practicing and code examples.
+#' This dataset was was generated on 2020-01-03 by running:
+#'
+#' \preformatted{
+#'   example_sensor <-
+#'     pat_createNew(
+#'       label = "SCAN_14",
+#'       pas = example_pas,
+#'       startdate = "2018-08-14",
+#'       enddate = "2018-09-07"
+#'     ) %>%
+#'     pat_createAirSensor(period = "1 hour")
+#'   save(example_sensor, file = "data/example_sensor.rda")
+#' }
+"example_sensor"
+
+
 #' #' @title Example AirSensor Timeseries dataset
 #' #' @format An S3 object composed of "meta" and "data" data.
 #' #' @description The \code{example_sensor_scaqmd} dataset provides a quickly